+++ python/openvip/models/message.py
@@ -70,6 +70,7 @@ class Message(BaseModel):
         validate_by_alias=True,
         validate_assignment=True,
         protected_namespaces=(),
+        extra="allow",
     )
 
 
@@ -123,17 +124,7 @@ class Message(BaseModel):
         if not isinstance(obj, dict):
             return cls.model_validate(obj)
 
-        _obj = cls.model_validate({
-            "openvip": obj.get("openvip"),
-            "type": obj.get("type"),
-            "id": obj.get("id"),
-            "timestamp": obj.get("timestamp"),
-            "text": obj.get("text"),
-            "origin": obj.get("origin"),
-            "language": obj.get("language"),
-            "trace_id": obj.get("trace_id"),
-            "parent_id": obj.get("parent_id")
-        })
+        _obj = cls.model_validate(obj)
         # store additional fields in additional_properties
         for _key in obj.keys():
             if _key not in cls.__properties: