+++ python/openvip/models/transcription.py
@@ -71,6 +71,7 @@ class Transcription(BaseModel):
         validate_by_alias=True,
         validate_assignment=True,
         protected_namespaces=(),
+        extra="allow",
     )
 
 
@@ -117,19 +118,7 @@ class Transcription(BaseModel):
         if not isinstance(obj, dict):
             return cls.model_validate(obj)
 
-        _obj = cls.model_validate({
-            "openvip": obj.get("openvip"),
-            "type": obj.get("type"),
-            "id": obj.get("id"),
-            "timestamp": obj.get("timestamp"),
-            "text": obj.get("text"),
-            "origin": obj.get("origin"),
-            "language": obj.get("language"),
-            "trace_id": obj.get("trace_id"),
-            "parent_id": obj.get("parent_id"),
-            "confidence": obj.get("confidence"),
-            "partial": obj.get("partial")
-        })
+        _obj = cls.model_validate(obj)
         return _obj